pyproject.toml
[project]
name = "kabosu-core"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-re2>=1.1.20250805",
    "jaconv>=0.4.0",
    "jpreprocess>=0.1.5",
    "kanalizer>=0.1.1",
    "marine-plus>=0.0.6.post3",
    "numpy==1.26.4",
    "onnxruntime>=1.22.1",
    "sudachidict-core>=20250825",
    "sudachipy>=0.6.10",
    "yomikata",
]

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
    "pytest"
]
ko = [
    "google-re2>=1.1.20250805",
    "jamo>=0.4.1",
    "mecab-ko>=1.0.2",
    "nltk>=3.9.1",
]


[tool.hatch.build.targets.sdist]
only-include = [
    "src/kabosu_core",
    "LICENSE",
    "pyproject.toml",
    "README.md",
]
[tool.hatch.build.targets.wheel]
packages = ["src/kabosu_core"]

[tool.uv.sources]
yomikata = { git = "https://github.com/q9uri/yomikata", rev = "ac5a27a2ed72ea4fab4e3a79e2d4799fb411e6bb" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
